Update advance.md
I consider myself new to Rust but I think there is an issue with this documentation....

The lifetime bound of a generic type, `T`, pertains to its lifetime as an entity and not the lifetime of the references it may contain. In particular, these references may be removed from it and preserved prior to it being dropped.

My reading of the rust nomenclature is as follows...

`T: 'a`   means `T` will live at least as long as `'a`

`'a: 'b`  means `'a' will live at least as long as `'b' and is therefore also substitutable for `'b' such as `'static : 'b`. The only value lifetimes have is their assertion of existence.

&'b T  is a reference to `T` that will not live longer that `'b`. The lifetime of, `'b` does not outlive whatever the lifetime of `T` is.